Implement container-based square div sizing
After "square" divs are added to the container, their size should be
equal to each other and spread out equally in the 960px x 960px
container.

Previous attempts were tried by using the "flex" css property but
haven't gave out the result needed. Thus, I decided that using the
Javascript to do the calculation of the squares and apply in-line CSS is
the temporary solution. The "flex" solution feels like a better solution
but I will have to look back into it.