Put code that create/remove the grid in functions
The site now has a button that allows the user to enter the number of
squares per side for their desired grid. The previous code only dealt
with 1 size (16x16). Now, the code is more flexible by being in a
function, allowing the program to easily make a grid in all sort of
sizes (cap to 100 in the future);

To make space for new squares to be generated, the old squares generated
from the old grid needs to be remove. Since we want this to run on every
button press, it's put inside a function for simplicity.