Create a function that returns correct grid size
When the user click on a button, the prompt will allow the user to type
in any value they could type on their keyboard. This needs to be
restricted because there's no way to produce a negative, decimal, or
string based grid size. Thus, this function is used to reduce the
chances that a weird error-inducing values will happen.