Remove background color in container & square divs
Previously, the background color was black for the container & pink for
the square divs. This was temporary placeholder to test visually if the
square divs were added correctly into the container.

Since all previous tests gave out expected results, these background
colors are removed & replaced with borders to make the container looks
more like a grid.